Revert "generalize country index creation method"
This reverts commit 9cf399e6d960325c661d5e7ae7bbd2e7e85663f7.

+++ lib/DB_Main.h
@@ -42,8 +42,8 @@ typedef struct Header
     uint off_job;           // Position bloc job
     uint off_ind;           // Position bloc industry
     uint off_grp;           // Position bloc group
-    long off_i_cty_name;    // Offset of the country index block (by name)
-    long i_cty_name;        // Address of the root for the country index (by name)
+    uint off_i_cty_name;    // Offset of the country index block (by name)
+    uint i_cty_name;        // Address of the root for the country index (by name)
     char filler[52];        // filler to get the size to 128 bytes
 } hder;
